Debugged the functions

+++ controllers/todoController.js
@@ -13,6 +13,13 @@ const addTodo = async (req, res) => {
             text: req.body.text,
             status: req.body.status ? req.body.status : false
         }
+        if (info.text === "") {
+            
+        return    res.status(404).send({
+                status: "fail",
+                message: "Cannot add empty string"
+            })
+        }
 
         const todo = await Todo.create(info)
         res.status(201).send(todo)
@@ -30,9 +37,14 @@ const getAllTodos = async (req, res) => {
 //READ a todo
 
 const getTodoById = async (req, res) => {
-    let id = req.params.id
-    let getATodo = await Todo.findOne({where: {id: id}})
-    res.status(200).send(getATodo)
+    let id = +req.params.id
+    if (id){
+        let getATodo = await Todo.findOne({where: {id: id}})
+        return res.status(200).send(getATodo)
+    } else {
+        return res.status(404).send({message: "couldn't find id"})
+    }
+        
 }
 
 
@@ -46,32 +58,32 @@ const completedTodo = async (req, res) => {
 //UPDATE
 
 const updateTodo = async (req, res) => {
-    let id = req.params.id
+    let id = +req.params.id
     Todo.findByPk(id).then(() => {
         info = {
             text: req.body.text,
         }
-        const todo = Todo.update(info, { where: { id } })
-        res.status(200).send({
+        const todo = Todo.update(info, { where: {id: id } })
+    return  res.status(200).send({
             message: "Updated successfully"
         })
     })
-            .catch((err) => {
-                console.lof(err)
-            });
+        .catch((err) => {
+            console.log(err)
+        });
 }
 
 
 //DELETE
 
 const deleteTodo = async (req, res) => {
-    let id = req.params.id
+    let id = +req.params.id
+    if (id)
+        await Todo.destroy({ where: { id: id } })
+    return  res.status(200).send({
 
-    await Todo.destroy({ where: { id: id } })
-    res.status(200).send({
-
-        message: "Todo successfully deleted"
-    })
+            message: "Todo successfully deleted"
+        })
 
 } 
 module.exports = {
